Track deployed indices
